requirements.txt
aiogram==2.25.1
aiogram-calendar==0.1.1
aiohttp==3.8.4
aioredis==1.3.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
docxcompose==1.4.0
docxtpl==0.16.7
dp==1.0
environs==9.5.0
frozenlist==1.3.3
hiredis==2.2.3
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.4
packaging==23.1
python-docx==0.8.11
python-dotenv==1.0.0
python-stdnum==1.18
pytz==2023.3
six==1.16.0
yarl==1.9.2
